+++ redisson-hibernate/redisson-hibernate-5/src/main/java/org/redisson/hibernate/RedissonRegionFactory.java
@@ -162,7 +162,7 @@ import java.util.concurrent.atomic.AtomicLong;
                             + "local nextValue = math.max(tonumber(ARGV[1]), tonumber(currentTime) + 1); "
                             + "redis.call('set', KEYS[1], nextValue); "
                             + "return nextValue;",
-                            RScript.ReturnType.INTEGER, Collections.singletonList("redisson-hibernate-timestamp"), time);
+                            RScript.ReturnType.LONG, Collections.singletonList("redisson-hibernate-timestamp"), time);
         } catch (Exception e) {
             if (fallback) {
                 while (true) {
