+++ redisson-hibernate/redisson-hibernate-53/src/main/java/org/redisson/hibernate/RedissonRegionFactory.java
@@ -163,7 +163,7 @@ public class RedissonRegionFactory extends RegionFactoryTemplate {
                             + "local nextValue = math.max(tonumber(ARGV[1]), tonumber(currentTime) + 1); "
                             + "redis.call('set', KEYS[1], nextValue); "
                             + "return nextValue;",
-                    RScript.ReturnType.INTEGER, Arrays.<Object>asList(qualifyName("redisson-hibernate-timestamp")), time);
+                    RScript.ReturnType.LONG, Arrays.<Object>asList(qualifyName("redisson-hibernate-timestamp")), time);
         } catch (Exception e) {
             if (fallback) {
                 return super.nextTimestamp();
